onchain_direction/requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# Bitcoin Price Direction Prediction — Q1 Paper
# Python 3.11+
#
# PyTorch must be installed separately with the correct CUDA version.
# For RTX 5060 (Blackwell, sm_120) run FIRST:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu128
# Then install the rest:
#   pip install -r requirements.txt
# ─────────────────────────────────────────────────────────────────────────────

# Core ML
scikit-learn==1.4.2
xgboost==2.0.3
shap==0.45.1

# Data
numpy==1.26.4
pandas==2.2.2
scipy==1.13.0

# Deep Learning (PyTorch installed separately — see above)
# torch>=2.6.0

# Statistical tests
statsmodels==0.14.2
pingouin==0.5.4

# Visualization
matplotlib==3.9.0
seaborn==0.13.2

# Utilities
pyyaml==6.0.1
tqdm==4.66.4
joblib==1.4.2
